requirements.txt
# ==================== Phase 1: 기본 챗봇 ====================

# FastAPI 웹 프레임워크
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
aiohttp==3.9.1

# Ollama Python 클라이언트
ollama==0.1.6

# Google Gemini API (무료 대안)
google-generativeai==0.3.2

# 이미지 처리 (기본)
Pillow==10.2.0
opencv-python==4.9.0.80
numpy==1.26.3

# PDF 생성 (견적서)
reportlab==4.0.9

# 데이터베이스
sqlalchemy==2.0.25
aiosqlite==0.19.0

# 유틸리티
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.31.0

# 카카오톡 API
flask==3.0.0

# 날짜/시간 처리
python-dateutil==2.8.2

# Rate Limiting (DDoS 방어)
slowapi==0.1.9

# 브라우저 자동화 (테스트/자동화)
playwright==1.55.0
pytest-playwright==0.7.1

# ==================== Phase 2: AI 이미지/영상 생성 (선택) ====================
# 아래 패키지는 Phase 2 고급 기능 사용 시에만 설치

# Transformers (Qwen2-VL)
# transformers>=4.37.0
# torch>=2.1.0
# torchvision>=0.16.0
# accelerate>=0.25.0

# ComfyUI는 별도 설치 필요
# git clone https://github.com/comfyanonymous/ComfyUI.git

# WebSocket (ComfyUI 통신)
# websocket-client==1.7.0
